Move splitting logic to its own file
