+++ test/core/login/keys.test.ts
@@ -1,11 +1,7 @@
 import { expect } from 'chai'
 import { describe, it } from 'mocha'
 
-import {
-  fixWalletInfo,
-  makeSplitWalletInfo,
-  mergeKeyInfos
-} from '../../../src/core/login/keys'
+import { fixWalletInfo, mergeKeyInfos } from '../../../src/core/login/keys'
 
 const ID_1 = 'PPptx6SBfwGXM+FZURMvYnsOfHpIKZBbqXTCbYmFd44='
 const ID_2 = 'y14MYFMP6vnip2hUBP7aqB6Ut0d4UNqHV9a/2vgE9eQ='
@@ -106,63 +102,6 @@ describe('fixWalletInfo', function () {
       id: 'id',
       keys: { bitcoinKey: 'bitcoinKey', format: 'bip44', coinType: 145 },
       type: 'wallet:bitcoin'
-    })
-  })
-})
-
-describe('splitWalletInfo', function () {
-  it('handles bitcoin to bitcoin cash', function () {
-    expect(
-      makeSplitWalletInfo(
-        fixWalletInfo({
-          id: 'MPo9EF5krFQNYkxn2I0elOc0XPbs2x7GWjSxtb5c1WU=',
-          type: 'wallet:bitcoin',
-          keys: {
-            bitcoinKey: '6p2cW62FeO1jQrbex/oTJ8R856bEnpZqPYxiRYV4fL8=',
-            dataKey: 'zm6w4Q0mNpeZJXrhYRoXiiV2xgONxvmq2df42/2M40A=',
-            syncKey: 'u8EIdKgxEG8j7buEt96Mq9usQ+k='
-          }
-        }),
-        'wallet:bitcoincash'
-      )
-    ).deep.equals({
-      id: 'SEsXNQxGL/D+8/vsBHJgwf7bAK6/OyR2BfescT7u/i4=',
-      type: 'wallet:bitcoincash',
-      keys: {
-        bitcoincashKey: '6p2cW62FeO1jQrbex/oTJ8R856bEnpZqPYxiRYV4fL8=',
-        dataKey: 'zm6w4Q0mNpeZJXrhYRoXiiV2xgONxvmq2df42/2M40A=',
-        syncKey: 'w3AiUfoTk8vQfAwPayHy/sJDH7E=',
-        format: 'bip32'
-      }
-    })
-  })
-
-  it('handles bitcoin cash to bitcoin', function () {
-    expect(
-      makeSplitWalletInfo(
-        {
-          id: 'MPo9EF5krFQNYkxn2I0elOc0XPbs2x7GWjSxtb5c1WU=',
-          type: 'wallet:bitcoincash',
-          keys: {
-            bitcoincashKey: '6p2cW62FeO1jQrbex/oTJ8R856bEnpZqPYxiRYV4fL8=',
-            dataKey: 'zm6w4Q0mNpeZJXrhYRoXiiV2xgONxvmq2df42/2M40A=',
-            syncKey: 'u8EIdKgxEG8j7buEt96Mq9usQ+k=',
-            format: 'bip44',
-            coinType: 145
-          }
-        },
-        'wallet:bitcoin'
-      )
-    ).deep.equals({
-      id: 'SEsXNQxGL/D+8/vsBHJgwf7bAK6/OyR2BfescT7u/i4=',
-      type: 'wallet:bitcoin',
-      keys: {
-        bitcoinKey: '6p2cW62FeO1jQrbex/oTJ8R856bEnpZqPYxiRYV4fL8=',
-        dataKey: 'zm6w4Q0mNpeZJXrhYRoXiiV2xgONxvmq2df42/2M40A=',
-        syncKey: 'w3AiUfoTk8vQfAwPayHy/sJDH7E=',
-        format: 'bip44',
-        coinType: 145
-      }
     })
   })
 })
